Backend-Api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
# SQLAlchemy - upgrade để tương thích với Python 3.13
sqlalchemy>=2.0.23
pymysql==1.1.0
# Cryptography - sử dụng version có pre-built wheels
cryptography>=41.0.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# Pydantic - sử dụng version mới hơn có pre-built wheels (không cần Rust)
pydantic>=2.5.0
pydantic-settings>=2.1.0
alembic==1.12.1
# Email
aiosmtplib==3.0.1
jinja2==3.1.2
email-validator==2.1.0

